+++ courseProgram/src/main/java/com/example/courseprogram/repository/TeacherRepository.java
@@ -7,12 +7,12 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TeacherRepository extends JpaRepository<Teacher,Integer> {
+public interface TeacherRepository extends JpaRepository<Teacher,Long> {
     @Modifying
     @Query(value = "update Teacher t set t = ?1")
     Teacher updateTeacherByTeacherId(Teacher teacher);
 
-    boolean existsTeacherByPerson_Number(Integer number);
+    boolean existsTeacherByPerson_Number(Long number);
 
     @Query(value = "select t from Teacher t,User u where u.userId=?1 and t.person.personId=u.person.personId")
     Teacher findByUserId(Integer userId);
